# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ForceFlow API — Multi-Tenant Salesforce Automation Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /onboarding/complete:
    post:
      operationId: completeOnboarding
      tags: [tenants]
      summary: Complete company onboarding wizard (creates tenant + connects first org + model)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "201":
          description: Onboarding complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /onboarding/invite/accept:
    post:
      operationId: acceptInvite
      tags: [tenants]
      summary: Accept a team invite and join a tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteAcceptInput"
      responses:
        "200":
          description: Invite accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantMember"
  /onboarding/invite/validate:
    get:
      operationId: validateInviteToken
      tags: [tenants]
      summary: Validate an invite token
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteDetails"

  # ─── TENANT ─────────────────────────────────────────────────────────────────
...
  /orgs:
    get:
      operationId: listOrgs
      tags: [orgs]
      summary: List all connected Salesforce orgs
      responses:
        "200":
          description: Orgs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesforceOrg"
    post:
      operationId: createOrg
      tags: [orgs]
      summary: Connect a new Salesforce org
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrgInput"
      responses:
        "201":
          description: Org connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesforceOrg"
  /orgs/{orgId}:
    get:
      operationId: getOrg
      tags: [orgs]
      summary: Get a specific org
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Org
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesforceOrg"
    delete:
      operationId: deleteOrg
      tags: [orgs]
      summary: Disconnect/delete an org
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /orgs/{orgId}/reconnect:
    post:
      operationId: reconnectOrg
      tags: [orgs]
      summary: Trigger OAuth reconnection for an org
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reconnect URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgReconnectResponse"

  # ─── MODEL CONFIGS ──────────────────────────────────────────────────────────
  /models:
    get:
      operationId: listModelConfigs
      tags: [models]
      summary: List all model configurations
      responses:
        "200":
          description: Model configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModelConfig"
    post:
      operationId: createModelConfig
      tags: [models]
      summary: Add a new model configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelConfigInput"
      responses:
        "201":
          description: Created model config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelConfig"
  /models/{modelId}:
    patch:
      operationId: updateModelConfig
      tags: [models]
      summary: Update a model configuration
      parameters:
        - name: modelId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelConfigUpdate"
      responses:
        "200":
          description: Updated model config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelConfig"
    delete:
      operationId: deleteModelConfig
      tags: [models]
      summary: Delete a model configuration
      parameters:
        - name: modelId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /models/{modelId}/test:
    post:
      operationId: testModelConfig
      tags: [models]
      summary: Test a model configuration connection
      parameters:
        - name: modelId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelTestResult"
  /models/{modelId}/set-default:
    post:
      operationId: setDefaultModel
      tags: [models]
      summary: Set a model config as the tenant default
      parameters:
        - name: modelId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated model config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelConfig"

  # ─── CHAT THREADS ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Tenant:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        companySize:
          type: ["string", "null"]
        sfClouds:
          type: array
          items:
            type: string
        orgDescription:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    TenantMember:
      type: object
      required: [id, tenantId, userId, role, joinedAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        userId:
          type: string
        email:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [frontier, developer, sales, support, sme]
        invitedBy:
          type: ["string", "null"]
        joinedAt:
          type: string
          format: date-time

    # ─── ONBOARDING ────────────────────────────────────────────────────────
    OnboardingInput:
      type: object
      required: [company]
      properties:
        company:
          type: object
          required: [name]
          properties:
            name:
              type: string
            website:
              type: string
            industry:
              type: string
            companySize:
              type: string
            sfClouds:
              type: array
              items:
                type: string
            orgDescription:
              type: string
        org:
          type: object
          properties:
            label:
              type: string
            loginUrl:
              type: string
            orgType:
              type: string
            clientId:
              type: string
            clientSecret:
              type: string
        model:
          type: object
          properties:
            provider:
              type: string
            apiKey:
              type: string
            modelName:
              type: string
            label:
              type: string
    InviteAcceptInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
        name:
          type: string
    InviteDetails:
      type: object
      required: [email, role, tenantName, tenantId, valid]
      properties:
        email:
          type: string
        role:
          type: string
        tenantName:
          type: string
        tenantId:
          type: string
        valid:
          type: boolean
        expiresAt:
          type: ["string", "null"]
          format: date-time

    # ─── TEAM ──────────────────────────────────────────────────────────────
...
    SalesforceOrg:
      type: object
      required: [id, tenantId, label, orgType, status, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        label:
          type: string
        instanceUrl:
          type: ["string", "null"]
        loginUrl:
          type: ["string", "null"]
        orgType:
          type: string
          enum: [production, sandbox, developer, scratch]
        username:
          type: ["string", "null"]
        status:
          type: string
          enum: [connected, disconnected, error]
        lastSync:
          type: ["string", "null"]
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    OrgInput:
      type: object
      required: [label, orgType, loginUrl, clientId, clientSecret]
      properties:
        label:
          type: string
        orgType:
          type: string
          enum: [production, sandbox, developer, scratch]
        loginUrl:
          type: string
        clientId:
          type: string
        clientSecret:
          type: string
    OrgReconnectResponse:
      type: object
      required: [authUrl]
      properties:
        authUrl:
          type: string

    # ─── MODEL CONFIGS ─────────────────────────────────────────────────────
    ModelConfig:
      type: object
      required: [id, tenantId, provider, modelName, isDefault, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        label:
          type: ["string", "null"]
        provider:
          type: string
          enum: [openai, anthropic, openrouter, custom]
        baseUrl:
          type: ["string", "null"]
        modelName:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ModelConfigInput:
      type: object
      required: [provider, modelName, apiKey]
      properties:
        label:
          type: string
        provider:
          type: string
          enum: [openai, anthropic, openrouter, custom]
        baseUrl:
          type: string
        modelName:
          type: string
        apiKey:
          type: string
        isDefault:
          type: boolean
    ModelConfigUpdate:
      type: object
      properties:
        label:
          type: string
        modelName:
          type: string
        apiKey:
          type: string
        baseUrl:
          type: string
        isDefault:
          type: boolean
    ModelTestResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        latencyMs:
          type: ["integer", "null"]

    # ─── CHAT ──────────────────────────────────────────────────────────────
...