# lib/api-zod/src/generated/types/modelConfigInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ForceFlow API — Multi-Tenant Salesforce Automation Platform
 * OpenAPI spec version: 0.1.0
 */
import type { ModelConfigInputProvider } from './modelConfigInputProvider';

export interface ModelConfigInput {
  label?: string;
  provider: ModelConfigInputProvider;
  baseUrl?: string;
  modelName: string;
  apiKey: string;
  isDefault?: boolean;
}