# lib/api-zod/src/generated/types/orgInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ForceFlow API — Multi-Tenant Salesforce Automation Platform
 * OpenAPI spec version: 0.1.0
 */
import type { OrgInputOrgType } from './orgInputOrgType';

export interface OrgInput {
  label: string;
  orgType: OrgInputOrgType;
  loginUrl: string;
  clientId: string;
  clientSecret: string;
}