# lib/api-zod/src/generated/types/tenantMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ForceFlow API — Multi-Tenant Salesforce Automation Platform
 * OpenAPI spec version: 0.1.0
 */
import type { TenantMemberRole } from './tenantMemberRole';

export interface TenantMember {
  id: string;
  tenantId: string;
  userId: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  name?: string | null;
  role: TenantMemberRole;
  /** @nullable */
  invitedBy?: string | null;
  joinedAt: Date;
}